extract filter trait
