+++ src/GraphQL/Queries/EntriesQuery.php
@@ -6,15 +6,15 @@ use GraphQL\Type\Definition\Type;
 use Statamic\Facades\Entry;
 use Statamic\Facades\GraphQL;
 use Statamic\GraphQL\Middleware\ResolvePage;
+use Statamic\GraphQL\Queries\Concerns\FiltersQuery;
 use Statamic\GraphQL\Types\EntryInterface;
 use Statamic\GraphQL\Types\JsonArgument;
-use Statamic\Support\Arr;
 use Statamic\Support\Str;
 use Statamic\Tags\Concerns\QueriesConditions;
 
 class EntriesQuery extends Query
 {
-    use QueriesConditions;
+    use QueriesConditions, FiltersQuery;
 
     protected $attributes = [
         'name' => 'entries',
@@ -53,31 +53,6 @@ class EntriesQuery extends Query
         $this->sortQuery($query, $args['sort'] ?? []);
 
         return $query->paginate($args['limit'] ?? 1000);
-    }
-
-    private function filterQuery($query, $filters)
-    {
-        if (! isset($filters['status']) && ! isset($filters['published'])) {
-            $filters['status'] = 'published';
-        }
-
-        foreach ($filters as $field => $definitions) {
-            if (! is_array($definitions)) {
-                $definitions = [['equals' => $definitions]];
-            }
-
-            if (Arr::assoc($definitions)) {
-                $definitions = collect($definitions)->map(function ($value, $key) {
-                    return [$key => $value];
-                })->values()->all();
-            }
-
-            foreach ($definitions as $definition) {
-                $condition = array_keys($definition)[0];
-                $value = array_values($definition)[0];
-                $this->queryCondition($query, $field, $condition, $value);
-            }
-        }
     }
 
     private function sortQuery($query, $sorts)
