+++ src/GraphQL/Queries/EntryQuery.php
@@ -5,14 +5,14 @@ namespace Statamic\GraphQL\Queries;
 use GraphQL\Type\Definition\Type;
 use Statamic\Facades;
 use Statamic\Facades\GraphQL;
+use Statamic\GraphQL\Queries\Concerns\FiltersQuery;
 use Statamic\GraphQL\Types\EntryInterface;
 use Statamic\GraphQL\Types\JsonArgument;
-use Statamic\Support\Arr;
 use Statamic\Tags\Concerns\QueriesConditions;
 
 class EntryQuery extends Query
 {
-    use QueriesConditions;
+    use QueriesConditions, FiltersQuery;
 
     protected $attributes = [
         'name' => 'entry',
@@ -62,30 +62,5 @@ class EntryQuery extends Query
         $this->filterQuery($query, $args['filter'] ?? []);
 
         return $query->limit(1)->get()->first();
-    }
-
-    private function filterQuery($query, $filters)
-    {
-        if (! isset($filters['status']) && ! isset($filters['published'])) {
-            $filters['status'] = 'published';
-        }
-
-        foreach ($filters as $field => $definitions) {
-            if (! is_array($definitions)) {
-                $definitions = [['equals' => $definitions]];
-            }
-
-            if (Arr::assoc($definitions)) {
-                $definitions = collect($definitions)->map(function ($value, $key) {
-                    return [$key => $value];
-                })->values()->all();
-            }
-
-            foreach ($definitions as $definition) {
-                $condition = array_keys($definition)[0];
-                $value = array_values($definition)[0];
-                $this->queryCondition($query, $field, $condition, $value);
-            }
-        }
     }
 }
